+++ .phan/plugins/RedundantAssignmentPlugin.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use ast\Node;
 use Phan\AST\UnionTypeVisitor;
+use Phan\Config;
 use Phan\Language\Context;
 use Phan\Language\Element\PassByReferenceVariable;
 use Phan\Parse\ParseVisitor;
@@ -103,6 +104,10 @@ class RedundantAssignmentPreAnalysisVisitor extends PluginAwarePreAnalysisVisito
         }
         if ($this->context->isInGlobalScope()) {
             if ($variable->getFileRef()->getFile() !== $this->context->getFile()) {
+                // Don't warn if this variable was set by a different file
+                return;
+            }
+            if (Config::getValue('__analyze_twice') && $variable->getFileRef()->getLineNumberStart() === $this->context->getLineNumberStart()) {
                 // Don't warn if this variable was set by a different file
                 return;
             }