+++ src/Phan/CLI.php
@@ -142,6 +142,7 @@ class CLI
         'language-server-tcp-server:',
         'language-server-verbose',
         'load-baseline:',
+        'analyze-twice',
         'long-progress-bar',
         'markdown-issue-messages',
         'memory-limit:',
@@ -828,6 +829,9 @@ class CLI
                     }
                     Config::setValue('baseline_path', $value);
                     break;
+                case 'analyze-twice':
+                    Config::setValue('__analyze_twice', true);
+                    break;
                 default:
                     // All of phan's long options are currently at least 2 characters long.
                     $key_repr = strlen($key) >= 2 ? "--$key" : "-$key";
@@ -1133,6 +1137,10 @@ class CLI
         if ($processes !== 1) {
             \fprintf(STDERR, "Notice: Running with processes=1 instead of processes=%s - the daemon/language server assumes it will run as a single process" . PHP_EOL, (string)\json_encode($processes));
             Config::setValue('processes', 1);
+        }
+        if (Config::getValue('__analyze_twice')) {
+            \fwrite(STDERR, "Notice: Running analysis phase once instead of --analyze-twice - the daemon/language server assumes it will run as a single process" . PHP_EOL);
+            Config::setValue('__analyze_twice', false);
         }
     }
 
@@ -1521,6 +1529,12 @@ $init_help
 
   (For best results, the baseline should be generated with the same/similar
   environment and settings as those used to run Phan)
+
+ --analyze-twice
+  Runs the analyze phase twice. Because Phan gathers additional type information for properties, return types, etc. during analysis,
+  this may emit a more complete list of issues.
+
+  This cannot be used with --processes <int>.
 
  -v, --version
   Print Phan's version number