+++ src/Phan/Config.php
@@ -913,6 +913,9 @@ class Config
         // E.g. this is used by `InvokePHPNativeSyntaxCheckPlugin`
         'plugin_config' => [
         ],
+
+        // This should only be set with `--analyze-twice`.
+        '__analyze_twice' => false,
     ];
 
     public const COMPLETION_VSCODE = 'vscode';
@@ -1258,6 +1261,7 @@ class Config
      *
      * @suppress PhanUnreferencedPublicMethod
      * @see FileRef::getProjectRelativePathForPath() for converting to relative paths
+     * NOTE: This deliberately does not support phar:// URLs, because those evaluate php code when the phar is first loaded.
      */
     public static function projectPath(string $relative_path): string
     {